+++ src/main/java/learn/thyme/thyteams/user/web/UserController.java
@@ -1,0 +1,14 @@
+package learn.thyme.thyteams.user.web;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/users")
+public class UserController {
+    @GetMapping
+    public String index() {
+        return "users/list";
+    }
+}